% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\docType{data}
\name{members}
\alias{members}
\title{Members of the Bundestag since 1949}
\format{
A tibble with 11618 observations from 4084 members of the Bundestag.
It includes 26 variables (in German):
\describe{
\item{\code{id}}{Identifikationsnummer}
\item{\code{nachname}}{Nachname}
\item{\code{vorname}}{Vorname}
\item{\code{adel}}{Adelsprädikat}
\item{\code{praefix}}{Namenspräfix}
\item{\code{anrede_titel}}{Anrede-Titel}
\item{\code{akad_titel}}{Akademischer Titel}
\item{\code{geburtsdatum}}{Geburtsdatum}
\item{\code{geburtsort}}{Geburtsort}
\item{\code{geburtsland}}{Geburtsland}
\item{\code{sterbedatum}}{Sterbedatum}
\item{\code{geschlecht}}{Geschlecht}
\item{\code{familienstand}}{Familienstand}
\item{\code{religion}}{Religion}
\item{\code{beruf}}{Beruf}
\item{\code{partei_kurz}}{Parteizugehörigkeit, kurzform}
\item{\code{vita_kurz}}{Kurzbiografie des Abgeordneten (nur aktuelle Wahlperiode)}
\item{\code{veroeffentlichungspflichtiges}}{Veröffentlichungspflichtige Angaben (nur aktuelle Wahlperiode)}
\item{\code{wp}}{Nummer der Wahlperiode}
\item{\code{mdbwp_von}}{Beginn der Wahlperiodenzugehörigkeit}
\item{\code{mdbwp_bis}}{Ende der Wahlperiodenzugehörigkeit}
\item{\code{wkr_nummer}}{Nummer des Wahlkreises}
\item{\code{wkr_name}}{Wahlkreisname}
\item{\code{wkr_land}}{Bundesland des Wahlkreises}
\item{\code{liste}}{Liste}
\item{\code{mandatsart}}{Art des Mandates}
}
}
\source{
Adapted from Bundestag (2020), \emph{Stammdaten aller Abgeordneten
seit 1949 im XML-Format}, version: 2020-03-12,
\url{https://www.bundestag.de/services/opendata}.
}
\usage{
members
}
\description{
A dataset with basic biographical and election data for all members of
the Bundestag since 1949.
}
\keyword{datasets}
