% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_available.R
\name{update_available}
\alias{update_available}
\title{Check for an updated version of the dataset on members of the Bundestag}
\usage{
update_available()
}
\value{
Returns TRUE if a more recent version of the data is available and FALSE
otherwise.
}
\description{
\code{update_available()} compares the version of the data loaded with the package
with the version currently available on the Bundestag website.
}
