% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_members.R
\name{import_members}
\alias{import_members}
\title{Import data on all members of the Bundestag since 1949}
\usage{
import_members()
}
\value{
A data frame of all parliamentary terms for all members
}
\description{
This function downloads the file "Stammdaten aller Abgeordneten seit 1949 im
XML-Format" from the Bundestag website converts it to a data frame, and
recodes some of the variables. The output is a longitudinal dataset, where
the level analysis is a member-term.
}
