% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\docType{data}
\name{members}
\alias{members}
\title{Members of the Bundestag since 1949.}
\format{
A tibble with 11618 observations from 4084 members of the Bundestag.
It includes 26 variables:
\describe{
\item{id}{A label}
\item{nachname}{A label}
\item{vorname}{A label}
\item{adel}{A label}
\item{praefix}{A label}
\item{anrede_titel}{A label}
\item{akad_titel}{A label}
\item{geburtsdatum}{A label}
\item{geburtsort}{A label}
\item{geburtsland}{A label}
\item{sterbedatum}{A label}
\item{geschlecht}{A label}
\item{familienstand}{A label}
\item{religion}{A label}
\item{beruf}{A label}
\item{partei_kurz}{A label}
\item{vita_kurz}{A label}
\item{veroeffentlichungspflichtiges}{A label}
\item{wp}{A label}
\item{mdbwp_von}{A label}
\item{mdbwp_bis}{A label}
\item{wkr_nummer}{A label}
\item{wkr_name}{A label}
\item{wkr_land}{A label}
\item{liste}{A label}
\item{mandatsart}{A label}
}
}
\source{
Adapted from Bundestag (2020), \emph{Stammdaten aller Abgeordneten
seit 1949 im XML-Format},
\url{https://www.bundestag.de/services/opendata}.
}
\usage{
members
}
\description{
A dataset with basic biographical and election data for all members of
the Bundestag since 1949.
}
\keyword{datasets}
